+++ raytracing/color.h
@@ -9,20 +9,20 @@ namespace cg
 class Color
 {
 public:
-    double r , g , b;
+    float r , g , b;
 
-    explicit Color( double R = 0 , double G = 0 , double B = 0 ) : r( R ) , g( G ) , b( B ) {}
+    explicit Color( float R = 0 , float G = 0 , float B = 0 ) : r( R ) , g( G ) , b( B ) {}
     ~Color() {}
 
     friend Color operator + ( const Color& , const Color& );
     friend Color operator - ( const Color& , const Color& );
     friend Color operator * ( const Color& , const Color& );
-    friend Color operator * ( const Color& , const double& );
-    friend Color operator / ( const Color& , const double& );
+    friend Color operator * ( const Color& , const float& );
+    friend Color operator / ( const Color& , const float& );
     friend Color& operator += ( Color& , const Color& );
     friend Color& operator -= ( Color& , const Color& );
-    friend Color& operator *= ( Color& , const double& );
-    friend Color& operator /= ( Color& , const double& );
+    friend Color& operator *= ( Color& , const float& );
+    friend Color& operator /= ( Color& , const float& );
     void Confine(); //luminance must be less than or equal to 1
     void Input( std::stringstream& );
 };
@@ -44,12 +44,12 @@ Color operator * ( const Color& A , const Color& B )
     return Color( A.r * B.r , A.g * B.g , A.b * B.b );
 }
 
-Color operator * ( const Color& A , const double& k )
+Color operator * ( const Color& A , const float& k )
 {
     return Color( A.r * k , A.g * k , A.b * k );
 }
 
-Color operator / ( const Color& A , const double& k )
+Color operator / ( const Color& A , const float& k )
 {
     return Color( A.r / k , A.g / k , A.b / k );
 }
@@ -70,7 +70,7 @@ Color& operator -= ( Color& A , const Color& B )
     return A;
 }
 
-Color& operator *= ( Color& A , const double& k )
+Color& operator *= ( Color& A , const float& k )
 {
     A.r *= k;
     A.g *= k;
@@ -78,7 +78,7 @@ Color& operator *= ( Color& A , const double& k )
     return A;
 }
 
-Color& operator /= ( Color& A , const double& k )
+Color& operator /= ( Color& A , const float& k )
 {
     A.r /= k;
     A.g /= k;
